Fix imports
follow-up to #98
